+++ packages/desktop/app/RoutingApp.js
@@ -1,5 +1,9 @@
 import React from 'react';
 import { Switch, Route, Redirect } from 'react-router';
+import { useSelector } from 'react-redux';
+
+import { SERVER_TYPES } from 'shared/constants';
+
 import { App } from './App';
 import {
   PatientsRoutes,
@@ -14,9 +18,16 @@ import {
   ReportsRoutes,
   ImmunisationRoutes,
 } from './routes';
+import { ConnectedFacilitySidebar, ConnectedSyncSidebar } from './components/Sidebar';
+import { TopBar, Notification } from './components';
 
-export const RoutingApp = React.memo(() => (
-  <App>
+export const RoutingApp = () => {
+  const isSyncServer = useSelector(state => state.auth?.server?.type === SERVER_TYPES.SYNC);
+  return isSyncServer ? <RoutingAdminApp /> : <RoutingFacilityApp />;
+};
+
+export const RoutingFacilityApp = React.memo(() => (
+  <App sidebarComponent={ConnectedFacilitySidebar}>
     <Switch>
       <Redirect exact path="/" to="/patients" />
       <Route path="/patients" component={PatientsRoutes} />
@@ -31,9 +42,31 @@ export const RoutingApp = React.memo(() => (
       <Route path="/reports" component={ReportsRoutes} />
       <Route path="/immunisations" component={ImmunisationRoutes} />
       {/*
-       * TODO fix this hack. For some reason, having an empty object within this switch fixes a bug
-       * where none of the app contents would render in a production build.
-       */}
+      * TODO fix this hack. For some reason, having an empty object within this switch fixes a bug
+      * where none of the app contents would render in a production build.
+      */}
     </Switch>
   </App>
 ));
+
+export const RoutingAdminApp = React.memo(() => (
+  <App sidebarComponent={ConnectedSyncSidebar}>
+    <Switch>
+      {/* <Redirect exact path="/" to="/admin" /> */}
+      <Route path="/" component={AdminPlaceholder} />
+      {/*
+      * TODO fix this hack. For some reason, having an empty object within this switch fixes a bug
+      * where none of the app contents would render in a production build.
+      */}
+    </Switch>
+  </App>
+));
+
+export const AdminPlaceholder = React.memo(() => {
+  const user = useSelector(state => state.auth?.user);
+
+  return <>
+    <TopBar title="New sync admin panel" />
+    <Notification message={`Successfully logged in as ${user.displayName}`} />
+  </>;
+});