+++ packages/desktop/app/api/TamanuApi.js
@@ -1,7 +1,7 @@
 import { promises } from 'fs';
 import qs from 'qs';
 
-import { VERSION_COMPATIBILITY_ERRORS } from 'shared/constants';
+import { VERSION_COMPATIBILITY_ERRORS, SERVER_TYPES } from 'shared/constants';
 import { LOCAL_STORAGE_KEYS } from '../constants';
 
 const { HOST, TOKEN, LOCALISATION, SERVER } = LOCAL_STORAGE_KEYS;
@@ -120,8 +120,14 @@ export class TamanuApi {
 
   async login(host, email, password) {
     this.setHost(host);
-    const response = await this.post('login', { email, password });
-    const { token, localisation, server, permissions } = response;
+    const response = await this.post('login', { email, password }, { returnResponse: true });
+    const serverType = response.headers.get('X-Tamanu-Server');
+    if (![SERVER_TYPES.LAN, SERVER_TYPES.SYNC].includes(serverType)) {
+      throw new Error(`Tamanu server type '${serverType}' is not supported.`);
+    }
+
+    const { token, localisation, server = {}, permissions } = await response.json();
+    server.type = serverType;
     saveToLocalStorage({ token, localisation, server });
     this.setToken(token);
     this.lastRefreshed = Date.now();
@@ -160,7 +166,7 @@ export class TamanuApi {
     if (!this.host) {
       throw new Error("TamanuApi can't be used until the host is set");
     }
-    const { headers, ...otherConfig } = config;
+    const { headers, returnResponse = false, ...otherConfig } = config;
     const queryString = qs.stringify(query || {});
     const url = `${this.prefix}/${endpoint}${query ? `?${queryString}` : ''}`;
     const response = await fetchOrThrowIfUnavailable(url, {
@@ -179,7 +185,11 @@ export class TamanuApi {
         this.refreshToken();
       }
 
-      return response.json();
+      if (returnResponse) {
+        return response;
+      } else {
+        return response.json();
+      }
     }
 
     const { error } = await getResponseJsonSafely(response);
@@ -203,11 +213,11 @@ export class TamanuApi {
     throw new Error(error?.message || response.status);
   }
 
-  async get(endpoint, query) {
-    return this.fetch(endpoint, query, { method: 'GET' });
+  async get(endpoint, query, options = {}) {
+    return this.fetch(endpoint, query, { method: 'GET', ...options });
   }
 
-  async postWithFileUpload(endpoint, filePath, body) {
+  async postWithFileUpload(endpoint, filePath, body, options = {}) {
     const fileData = await promises.readFile(filePath);
     const blob = new Blob([fileData]);
 
@@ -223,30 +233,33 @@ export class TamanuApi {
     return this.fetch(endpoint, null, {
       method: 'POST',
       body: formData,
+      ...options,
     });
   }
 
-  async post(endpoint, body) {
+  async post(endpoint, body, options = {}) {
     return this.fetch(endpoint, null, {
       method: 'POST',
       body: body && JSON.stringify(body),
       headers: {
         'Content-Type': 'application/json',
       },
+      ...options,
     });
   }
 
-  async put(endpoint, body) {
+  async put(endpoint, body, options = {}) {
     return this.fetch(endpoint, null, {
       method: 'PUT',
       body: body && JSON.stringify(body),
       headers: {
         'Content-Type': 'application/json',
       },
+      ...options,
     });
   }
 
-  async delete(endpoint, query) {
-    return this.fetch(endpoint, query, { method: 'DELETE' });
+  async delete(endpoint, query, options = {}) {
+    return this.fetch(endpoint, query, { method: 'DELETE', ...options });
   }
 }