+++ packages/sync-server/app/middleware/versionCompatibility.js
@@ -10,6 +10,10 @@ export const SUPPORTED_CLIENT_VERSIONS = {
     min: '1.16.0',
     max: '1.16.0', // note that higher patch versions will be allowed to connect
   },
+  'Tamanu Desktop': {
+    min: '1.16.0',
+    max: '1.16.0', // note that higher patch versions will be allowed to connect
+  },
   'Tamanu Mobile': {
     min: '1.16.0',
     max: '1.16.99', // note that higher patch versions will be allowed to connect
@@ -25,8 +29,7 @@ export const SUPPORTED_CLIENT_VERSIONS = {
 };
 
 export const versionCompatibility = (req, res, next) => {
-  // TODO: X-Runtime is deprecated
-  const clientType = req.header('X-Tamanu-Client') || req.header('X-Runtime');
+  const clientType = req.header('X-Tamanu-Client');
 
   if (!clientType) {
     // a thirdparty tool (or internal test suite) is using the API; ignore version checking